Moves `fetchRegisteredSteamAPIKey` to Separate File
